MonteCarloBot move
